Fix unrelated dispatches of messages
When handling events in batch, dispatching of messages was possibly recorded on the wrong message. Messages were recorde on the afterCommit, at which point the handlerInformation of the unit of work had been set.

This PR removed caching of the handlerInformation and moves recording of dispatched messages to the span.