Baseline to 4.6.x and fix SpanFactory
The spanfactory.makeCurrent did not always work, since the scope can differ. We implemented makeCurrent because we thought we should work similarly to OpenTelemetry, but this is not the case. The span's scope is ended too early, and the actual handler identification is not reported.

The HandlerEnhancerDefinition has been moved back to serviceloader mechanism, so we can baseline to 4.6.x