Detect decorated components and unwrap them
When using Inspector Axon, and components are decorated, this leads to inaccurate information displayed in the UI.
For example, when using FireStarter and Axon Server, the UI will show that the customer is not using Axon Server, because FireStartEventStore is not the AxonServerEventStore. By unwrapping it, we can show the correct information.

As a later feature we might consider showing the whole decoration chain.