Fix issue where configuration ordering caused problems for metrics.
Because the postprocessor depended on the entire Configuration object of Axon, it caused the MetricRegistry to initialize early. This in turn caused Spring Boot metrics to be missing from the metric registry.

The HandlerMetricsRegistry now sets a static instance upon start, which is picked up by the SpanFactory. This removed any dependency, causing normal initialization order again.

Fixes #33