Fix latency while batching is active, and remove latency from processor level.
This commit changes when the commit latency is registered. When batching, the latency was measured of the first event in the batch, due to the reverse callback order of the afterCommit hook. The afterCommit hook is now only set for the last message, so the latency of the last message in the batch is measured. This was only noticable during very few events and while replaying.

In addition the latencies were removed from the processor level of the report. These were unusable, because latency for unclaimed segments was kept. The IA server should take care of determining how to handle individual latency values of segments and aggregate them, not the client.