Fix rolling count to be more reliable measurement
This PR fixes a predicament with the RollingCountMeasure, which often led to missing invocations of a message handler due to the way the values rotates.
The variables have been replaced by a map, and counts are separated into buckets of 2.5 seconds. The value is calculated over the time in the last minute before the last 2.5 seconds began (as to not include incomplete buckets and provide inaccurate values).