Fix problem with SagaScopeDescriptor
If a command was dispatched on the local command bus by a saga, th saga scope would be active and a classcast exception could occur. This has been resolved by adding a type check.
In addition, the entire method has been try-catched in order to prevent interruptions in the future.

Last but not least, some API properties were re-added as deprecated so the connector can be used by people before the new server is running.