Add the CLAIM operation for event processors
This operation will always wait until the segment is claimed by the processor, up to 30 seconds.
If the target is a PSEP, it will act actively by setting the `unclaimedSegmentValidationThreshold` of the `CoordinationTask` to `0` and scheduling it immediately. In my tests this will happen in <100ms, preventing 'downtime' of segments during scaling, merge and split operations.