Optimize CLAIM functionality pre- and post-4.9
Adds an optimization for both TEPs and PSEPs to clear the release deadline for a segment upon a CLAIM operation.

Also added warnings  when people are on a version lower than 4.9.x, and pre-added the integration code necessary for that future version so people do not need to upgrade the inspector-axon dependency when it comes out.